Moved network_session stuff to separate files
fixed compiling when adding header files to GUI.cpp
