--- xlive/H2MOD/Modules/GameManager/GameManager.h
@@ -1,12 +1,0 @@
-#pragma once
-
-class GameManager {
-public:
-	bool isHost(); 
-	bool start(); 
-private:
-	bool started = false;
-
-};
-
-extern GameManager* gameManager;
